requirements.txt
snowflake-connector-python
python-dotenv
requests
dbt-snowflake
streamlit
pandas
altair